Reorganise Interprereter code + new functions
